Refactor the collection logic
