+++ pkg/collector/corechecks/cluster/orchestrator/orchestrator_processor.go
@@ -394,67 +394,6 @@ func chunkReplicaSets(replicaSets []*model.ReplicaSet, chunkCount, chunkSize int
 	return chunks
 }
 
-// processServiceList process a service list into process messages.
-func processServiceList(serviceList []*corev1.Service, groupID int32, cfg *config.OrchestratorConfig, clusterID string) ([]model.MessageBody, error) {
-	start := time.Now()
-	serviceMsgs := make([]*model.Service, 0, len(serviceList))
-
-	for s := 0; s < len(serviceList); s++ {
-		svc := serviceList[s]
-		if orchestrator.SkipKubernetesResource(svc.UID, svc.ResourceVersion, orchestrator.K8sService) {
-			continue
-		}
-		redact.RemoveLastAppliedConfigurationAnnotation(svc.Annotations)
-
-		serviceModel := extractService(svc)
-
-		// k8s objects only have json "omitempty" annotations
-		// + marshalling is more performant than YAML
-		jsonSvc, err := jsoniter.Marshal(svc)
-		if err != nil {
-			log.Warnf("Could not marshal Service to JSON: %s", err)
-			continue
-		}
-		serviceModel.Yaml = jsonSvc
-
-		serviceMsgs = append(serviceMsgs, serviceModel)
-	}
-
-	groupSize := orchestrator.GroupSize(len(serviceMsgs), cfg.MaxPerMessage)
-
-	chunks := chunkServices(serviceMsgs, groupSize, cfg.MaxPerMessage)
-	messages := make([]model.MessageBody, 0, groupSize)
-
-	for i := 0; i < groupSize; i++ {
-		messages = append(messages, &model.CollectorService{
-			ClusterName: cfg.KubeClusterName,
-			ClusterId:   clusterID,
-			GroupId:     groupID,
-			GroupSize:   int32(groupSize),
-			Services:    chunks[i],
-			Tags:        cfg.ExtraTags,
-		})
-	}
-
-	orchestrator.SetCacheStats(len(serviceList), len(serviceMsgs), orchestrator.K8sService)
-
-	log.Debugf("Collected & enriched %d out of %d Services in %s", len(serviceMsgs), len(serviceList), time.Since(start))
-	return messages, nil
-}
-
-// chunkServices chunks the given list of services, honoring the given chunk count and size.
-// The last chunk may be smaller than the others.
-func chunkServices(services []*model.Service, chunkCount, chunkSize int) [][]*model.Service {
-	chunks := make([][]*model.Service, 0, chunkCount)
-
-	for counter := 1; counter <= chunkCount; counter++ {
-		chunkStart, chunkEnd := orchestrator.ChunkRange(len(services), chunkCount, chunkSize, counter)
-		chunks = append(chunks, services[chunkStart:chunkEnd])
-	}
-
-	return chunks
-}
-
 func fillClusterResourceVersion(c *model.Cluster) error {
 	// Marshal the cluster message to JSON.
 	marshaller := jsoniter.ConfigCompatibleWithStandardLibrary
